backend/requirements.txt
# ==============================================================
# Django & REST Framework
# ==============================================================
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0

# ==============================================================
# Database
# ==============================================================
# SQLite is built-in; uncomment the line below if using MySQL:
# mysqlclient==2.2.7

# ==============================================================
# Face Recognition (requires dlib + cmake)
# ==============================================================
# NOTE: dlib needs CMake and a C++ compiler installed on your system.
#   Ubuntu/Debian:  sudo apt install cmake build-essential
#   macOS:          brew install cmake
cmake==4.2.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0

# ==============================================================
# Image handling
# ==============================================================
pillow==12.1.0
numpy==2.4.1

# ==============================================================
# Utilities
# ==============================================================
asgiref==3.11.0
sqlparse==0.5.5
setuptools==80.10.2
click==8.3.1
